requirements.txt
aiohttp                   3.9.5
aiosignal                 1.3.1
altair                    5.3.0
annotated-types           0.6.0
anyio                     4.3.0
async-timeout             4.0.3
attrs                     23.2.0
blinker                   1.7.0
cachetools                5.3.3
certifi                   2024.2.2
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
dataclasses-json          0.6.4
distro                    1.9.0
docarray                  0.40.0
dotenv                    0.0.5
exceptiongroup            1.2.1
frozenlist                1.4.1
gitdb                     4.0.11
GitPython                 3.1.43
greenlet                  3.0.3
h11                       0.14.0
hnswlib                   0.8.0
httpcore                  1.0.5
httpx                     0.27.0
idna                      3.7
Jinja2                    3.1.3
jsonpatch                 1.33
jsonpointer               2.4
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
langchain                 0.1.16
langchain-community       0.0.34
langchain-core            0.1.45
langchain-text-splitters  0.0.1
langsmith                 0.1.49
markdown-it-py            3.0.0
MarkupSafe                2.1.5
marshmallow               3.21.1
mdurl                     0.1.2
multidict                 6.0.5
mypy-extensions           1.0.0
numpy                     1.26.4
openai                    1.23.2
orjson                    3.10.1
packaging                 23.2
pandas                    2.2.2
pillow                    10.3.0
pip                       24.0
protobuf                  4.25.3
pyarrow                   16.0.0
pydantic                  2.7.0
pydantic_core             2.18.1
pydeck                    0.8.1b0
Pygments                  2.17.2
pypdf                     4.2.0
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
pytz                      2024.1
PyYAML                    6.0.1
referencing               0.34.0
regex                     2024.4.16
requests                  2.31.0
rich                      13.7.1
rpds-py                   0.18.0
setuptools                57.4.0
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.1
SQLAlchemy                2.0.29
streamlit                 1.33.0
tenacity                  8.2.3
tiktoken                  0.6.0
toml                      0.10.2
toolz                     0.12.1
tornado                   6.4
tqdm                      4.66.2
types-requests            2.31.0.20240406
typing_extensions         4.11.0
typing-inspect            0.9.0
tzdata                    2024.1
urllib3                   2.2.1
watchdog                  4.0.0
yarl                      1.9.4